wip: vici: Make per-CPU CHILD_SAs configurable
wip: maybe we could just make this a boolean and determine the number of
CPUs automatically?  or does it make sense to set this to a lower value
(we can't enforce that on the policy level)?
also, maybe add a check to ensure that start_action=trap is set (or even
set it implicitly?), but could also just be the responsibility of the
user as it's documented